requirements.txt
absl-py==0.11.0
-e git+https://gitee.com/xieyulai/alphapose_for_action.git@c3f589406ed774e9958374f24dd4a8b4198e0257#egg=alphapose
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
easydict==1.9
einops==0.3.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
h5py==3.2.1
idna==2.10
imageio==2.13.5
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
munkres==1.1.4
natsort==8.1.0
networkx==2.6.3
numpy==1.19.4
oauthlib==3.1.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
ordered-set==4.0.2
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.0
pycocotools==2.0.0
pygame==2.1.2
pyheatmap==0.1.12
pynvml==11.0.0
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-fast-transformers==0.4.0
pytz==2020.4
PyWavelets==1.2.0
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
timm==0.1.20
toml==0.10.2
torch==1.10.1
torchstat==0.0.7
torchsummaryX==1.3.0
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.2
visdom==0.1.8.9
Werkzeug==1.0.1
yacs==0.1.8
